+++ PEPPROr/R/PEPPROr.R
@@ -41,25 +41,25 @@ NULL
 #' @examples
 #' data("ccurve")
 #' data("counts")
-#' plot_complexity_curves(ccurve, coverage=3099922541, read_length=30,
-#'                        real_counts_path=counts)
-plot_complexity_curves <- function(ccurves,
-                                   coverage=0, read_length=0,
-                                   real_counts_path=NA, ignore_unique=FALSE,
-                                   output_name='complexity_curves',
-                                   x_min=0, x_max=500000000) {
+#' plotComplexityCurves(ccurves = "ccurve", coverage=3099922541, read_length=30,
+#'                      real_counts_path="counts")
+plotComplexityCurves <- function(ccurves,
+                                 coverage=0, read_length=0,
+                                 real_counts_path=NA, ignore_unique=FALSE,
+                                 output_name='complexity_curves',
+                                 x_min=0, x_max=500000000) {
 
     if (x_min < 0 || x_max <= x_min) {
         message(paste0("problem with x-min or x-max (", x_min, " ", x_max,
                        "). x-min must be >= 0 and < x-max"))
-        quit()
+        quit(save = "no", status = 1, runLast = FALSE)
     }
 
     # Convert limit counts to coverage
     if (coverage > 0) {
         if (read_length == 0) {
             message("Error: --coverage specified but not --read_length")
-            quit()
+            quit(save = "no", status = 1, runLast = FALSE)
         } else {
             coverage <- as.numeric(coverage) / as.numeric(read_length)
         }
@@ -71,8 +71,9 @@ plot_complexity_curves <- function(ccurves,
     rcDT <- data.table(name = character(length(real_counts_path)),
                        total = integer(length(real_counts_path)),
                        unique = integer(length(real_counts_path)))
-    if ("data.frame" %in% class(real_counts_path[1])) {
-        rc_file    <- real_counts_path[1]
+
+    if (exists(real_counts_path[1])) {
+        rc_file    <- data.table(get(real_counts_path[1]))
         rcDT$name  <- basename(rc_file$V1)
         rcDT$total <- as.integer(rc_file$V2)
         if (ncol(rc_file) == 3 && !ignore_unique) {
@@ -113,7 +114,7 @@ plot_complexity_curves <- function(ccurves,
                 } else if (info$size == 0) {
                     message("File is empty.")
                 }
-                quit()
+                quit(save = "no", status = 1, runLast = FALSE)
             }
         }
     }
@@ -144,7 +145,17 @@ plot_complexity_curves <- function(ccurves,
         for(j in 1:numFields) name <- gsub("_[^_]*$", "", name)
         sample_name <- name
         message(paste0("Processing ", sample_name))
-        ctable <- fread(ccurves[[c]])
+
+        if (exists(ccurves[[c]])) {
+            ctable <- data.table(get(ccurves[[c]]))
+        } else if (file.exists(ccurves[[c]])) {
+            ctable <- fread(ccurves[[c]])
+        } else {
+            stop(paste0("FileExistsError: ", ccurves[[c]],
+                        " could not be found."))
+            quit(save = "no", status = 1, runLast = FALSE)
+        }
+
         if (coverage > 0) {
             if ("TOTAL_READS" %in% colnames(ctable)) {
                 ccurve_TOTAL_READS <- as.numeric(ctable$TOTAL_READS) / coverage
@@ -160,7 +171,7 @@ plot_complexity_curves <- function(ccurves,
             } else {
                 messsage(paste0("Error, table ", c, " is not in the expected "))
                 message("format... has it been generated with preseq?")
-                quit()
+                quit(save = "no", status = 1, runLast = FALSE)
             }
         } else {
             if ("TOTAL_READS" %in% colnames(ctable)) {
@@ -175,7 +186,7 @@ plot_complexity_curves <- function(ccurves,
             } else {
                 messsage(paste0("Error, table ", c, " is not in the expected "))
                 message("format... has it been generated with preseq?")
-                quit()
+                quit(save = "no", status = 1, runLast = FALSE)
             }
         }
         if (c == 1) {
@@ -463,9 +474,11 @@ computeLimit <- function(value, ccurve_TOTAL_READS) {
     last_point   <- length(ccurve_TOTAL_READS)
     iterations   <- 0
     while (first_point != last_point) {
-        middle_point <- (first_point + last_point)/2
-        middle_value <- ccurve_TOTAL_READS[middle_point]
-        if (middle_value == value || iterations >= 10000) {
+        middle_point <- as.numeric((first_point + last_point)/2)
+        middle_value <- as.numeric(ccurve_TOTAL_READS[middle_point])
+        if (length(middle_value)==0) {
+            return(middle_point)
+        } else if (middle_value == value || iterations >= 10000) {
             return(middle_point)
         } else if (middle_value >= value) {
             last_point  <- middle_point - 1
@@ -596,7 +609,7 @@ plotFRiF <- function(sample_name, num_reads, output_name, bedFile) {
             } else {
                 outFile <- file.path(output_name)
                 system2(paste("touch"), outFile)
-                quit()
+                quit(save = "no", status = 1, runLast = FALSE)
             }
 
             if (max(bed[,4] > 0)) {
@@ -668,6 +681,16 @@ plotFRiF <- function(sample_name, num_reads, output_name, bedFile) {
 }
 
 
+#' The function rounds the up to the nearest "nice" number.
+#'
+#' From:
+#' https://stackoverflow.com/questions/6461209/how-to-round-up-to-the-nearest-10-or-100-or-x
+roundUpNice <- function(x, nice=c(1,2,3,4,5,6,7,8,9,10)) {
+    if(length(x) != 1) stop("'x' must be of length 1")
+    10^floor(log10(x)) * nice[[which(x <= 10^floor(log10(x)) * nice)[[1]]]]
+}
+
+
 #' Plot TSS enrichment
 #'
 #' This function plots the global TSS enrichment and produces pdf/png files.
@@ -677,7 +700,7 @@ plotFRiF <- function(sample_name, num_reads, output_name, bedFile) {
 #' @export
 #' @examples
 #' data("tss")
-#' plotTSS(TSSfile = "TSSfile")
+#' plotTSS(TSSfile = "tss")
 #' @export
 plotTSS <- function(TSSfile) {
     if (length(TSSfile) == 1) {
@@ -692,7 +715,7 @@ plotTSS <- function(TSSfile) {
                          paste(TSSfile, collapse=", ")),
                   stdout())
             write(paste0("Did you mean to pass more than 2 files?"), stdout())
-            quit()
+            quit(save = "no", status = 1, runLast = FALSE)
         }
     }
 
@@ -719,7 +742,7 @@ plotTSS <- function(TSSfile) {
 
     iMat <- data.table(V1 = numeric())
     if (length(TSSfile) == 1) {
-        if (exists(TSSfile[i])) {
+        if (exists(TSSfile)) {
             iMat <- data.table(get(TSSfile))
         } else {
             iMat <- fread(TSSfile)
@@ -745,7 +768,7 @@ plotTSS <- function(TSSfile) {
                      paste(TSSfile, collapse=", ")),
               stdout())
         write(paste0("Did you mean to pass more than 2 files?"), stdout())
-        quit()
+        quit(save = "no", status = 1, runLast = FALSE)
     }
 
     if (length(TSSfile) == 1) {
@@ -766,7 +789,7 @@ plotTSS <- function(TSSfile) {
                                        peakPos+50)), score]),1)
         if (is.nan(TSSscore)) {
             message(paste0("\nNaN produced.  Check ", TSSfile, "\n"))
-            quit()
+            quit(save = "no", status = 1, runLast = FALSE)
         }
     } else {
         val      <- 0.05*nrow(plus)
@@ -779,7 +802,7 @@ plotTSS <- function(TSSfile) {
                                        peakPos+50)), score]),1)
         if (is.nan(TSSscore)) {
             message(paste0("\nNaN produced.  Check ", TSSfile[1], "\n"))
-            quit()
+            quit(save = "no", status = 1, runLast = FALSE)
         }
     }
     
@@ -802,10 +825,11 @@ plotTSS <- function(TSSfile) {
         geom_smooth(method="loess", span=0.02,
                     se=FALSE, colour=lineColor) +
         labs(x = "Distance from TSS (bp)", y = "TSS Enrichment Score")
+    y_max <- max(30, roundUpNice(TSSscore))
     p <- pre + t1 +
          scale_x_continuous(expand=c(0,0)) +
          scale_y_continuous(expand=c(0,0)) +
-         coord_cartesian(xlim=c(-2300,2300), ylim=c(0,32))
+         coord_cartesian(xlim=c(-2300, 2300), ylim=c(0, y_max+2))
     if (exists("minus")) {
         val      <- 0.025*nrow(minus)
         # normTSS  <- (minus / mean(minus[c(1:val,
@@ -818,7 +842,7 @@ plotTSS <- function(TSSfile) {
                                peakPos+50)), score]),1)
         if (is.nan(minusTSSscore)) {
             message(paste0("\nNaN produced.  Check ", TSSfile[2], "\n"))
-            quit()
+            quit(save = "no", status = 1, runLast = FALSE)
         }
         p <- p + geom_smooth(data=minusNormTSS,
                              aes(x=(as.numeric(rownames(minusNormTSS))-
@@ -826,25 +850,25 @@ plotTSS <- function(TSSfile) {
                                  y=score, group=1, colour="black"),
                              method="loess", span=0.02,
                              se=FALSE, colour="blue") +
-                 annotate("rect", xmin=1200, xmax=2300, ymin=25, ymax=32,
-                          fill="gray95", size = 0.5) +
-                 annotate("text", x=1750, y=31, label="TSS Score", fontface = 1,
-                          size=6, hjust=0.5) +
-                 annotate("text", x=1500, y=29, label="+", fontface = 2,
+                 annotate("rect", xmin=1200, xmax=2300, ymin=y_max-8,
+                          ymax=y_max+2, fill="gray95", size = 0.5) +
+                 annotate("text", x=1750, y=y_max, label="TSS Score",
+                          fontface = 1, size=6, hjust=0.5) +
+                 annotate("text", x=1500, y=y_max-2, label="+", fontface = 2,
                           size=8, hjust=0.5, color=lineColor) +
-                 annotate("text", x=1500, y=27, label=TSSscore, fontface = 2,
-                          size=8, hjust=0.5, color=lineColor) +
-                 annotate("text", x=2000, y=29, label="-",
+                 annotate("text", x=1500, y=y_max-5, label=TSSscore,
+                          fontface = 2, size=8, hjust=0.5, color=lineColor) +
+                 annotate("text", x=2000, y=y_max-2, label="-",
                           fontface = 2, size=8, hjust=0.5, color="blue") +
-                 annotate("text", x=2000, y=27, label=minusTSSscore,
+                 annotate("text", x=2000, y=y_max-5, label=minusTSSscore,
                           fontface = 2, size=8, hjust=0.5, color="blue")
     } else {
-        p <- p + annotate("rect", xmin=1200, xmax=2300, ymin=27, ymax=32,
-                          fill="gray95", size = 0.5) +
-                 annotate("text", x=1750, y=31, label="TSS Score",
+        p <- p + annotate("rect", xmin=1200, xmax=2300, ymin=y_max-4,
+                          ymax=y_max+2, fill="gray95", size = 0.5) +
+                 annotate("text", x=1750, y=y_max+1, label="TSS Score",
                           fontface = 1, size=6, hjust=0.5) +
-                 annotate("text", x=1750, y=29, label=TSSscore, fontface = 2,
-                          size=10, hjust=0.5)
+                 annotate("text", x=1750, y=y_max-1, label=TSSscore,
+                          fontface = 2, size=10, hjust=0.5)
     }
 
     png(filename = paste0(sample_name, "_TSSenrichment.png"),
@@ -873,9 +897,9 @@ plotTSS <- function(TSSfile) {
 #' @keywords fragment distribution
 #' @export
 #' @examples
-#' data("fragL")
-#' data("fragL_count")
-#' plotFLD(fragL = "fragL", fragL_count = "fragL_count",
+#' data("frag_len")
+#' data("frag_len_count")
+#' plotFLD(fragL = "frag_len", fragL_count = "frag_len_count",
 #'         fragL_dis1 = "fragLenDistribution_example.pdf",
 #'         fragL_dis2 = "fragLenDistribution_example.txt")
 #' @export
@@ -883,23 +907,37 @@ plotFLD <- function(fragL, fragL_count,
                     fragL_dis1="fragLenDistribution.pdf",
                     fragL_dis2="fragLenDistribution.txt") {
 
-    outfile_png <- gsub('pdf', 'png', fragL_dis1)
+    if (exists(fragL_count)) {
+        dat <- data.table(get(fragL_count))
+    } else if (file.exists(fragL_count)) {
+        dat <- fread(fragL_count)
+    } else {
+        stop(paste0("FileExistsError: ", fragL_count, " could not be found."))
+        quit(save = "no", status = 1, runLast = FALSE)
+    }
 
-    dat  <- fread(fragL_count)
+    if (exists(fragL)) {
+        summary_table <- data.table(get(fragL))
+    } else if (file.exists(fragL)) {
+        summary_table <- fread(fragL)
+    } else {
+        stop(paste0("FileExistsError: ", fragL, " could not be found."))
+        quit(save = "no", status = 1, runLast = FALSE)
+    }
+
     dat1 <- dat[dat$V2<=600,]
     tmp  <- seq(1:as.numeric(dat1[1,2]-1))
     dat0 <- data.table(V1=rep(0,length(tmp)),V2=tmp)
     dat2 <- rbind(dat0, dat1)
 
     t1 = theme_classic(base_size=14) +
-         theme(axis.line = element_line(size = 0.5)) +
-         theme(panel.grid.major = element_blank(),
+         theme(axis.line = element_line(size = 0.5), 
+               panel.grid.major = element_blank(),
                panel.grid.minor = element_blank(),
                legend.position = "none",
                aspect.ratio = 1,
-               panel.border = element_rect(colour = "black",
-                                          fill=NA, size=0.5)) +
-         theme(plot.title = element_text(hjust = 0.5))
+               panel.border = element_rect(colour = "black", fill=NA, size=0.5),
+               plot.title = element_text(hjust = 0.5))
 
     p <- ggplot(dat1, aes(x=V2, y=V1)) +
              geom_line(aes(color='red')) +
@@ -912,15 +950,18 @@ plotFLD <- function(fragL, fragL_count,
     pdf(file=fragL_dis1, width= 7, height = 7, useDingbats=F)
     print(p)
     invisible(dev.off())
-         
+
     # Save plot to png file
+    outfile_png <- gsub('pdf', 'png', fragL_dis1)
     png(filename=outfile_png, width = 480, height = 480)
     print(p)
     invisible(dev.off())
 
-    dat  <- fread(fragL)
-    summ <- data.table(Min=min(dat$V1), Max=max(dat$V1), Median=median(dat$V1),
-                       Mean=mean(dat$V1), Stdev=sd(dat$V1))
+    summ <- data.table(Min=min(summary_table$V1),
+                       Max=max(summary_table$V1),
+                       Median=median(summary_table$V1),
+                       Mean=mean(summary_table$V1),
+                       Stdev=sd(summary_table$V1))
     # Write summary table to stats file
     fwrite(summ, file=fragL_dis2, row.names=F, quote=F, sep="\t")
 }
@@ -1001,14 +1042,17 @@ fancyNumbers <- function(n){
 #' @keywords mRNA contamination
 #' @export
 #' @examples
-#' data("rpkm")
-#' mRNAcontamination(rpkm = "rpkm")
+#' data("rpkm_ratios")
+#' mRNAcontamination(rpkm = "rpkm_ratios")
 #' @export
 mRNAcontamination <- function(rpkm, raw=FALSE) {
     if (exists(rpkm)) {
         RPKM <- data.table(get(rpkm))
-    } else {
+    } else if (file.exists(rpkm)) {
         RPKM <- fread(rpkm)
+    } else {
+        stop(paste0("FileExistsError: ", rpkm, " could not be found."))
+        quit(save = "no", status = 1, runLast = FALSE)
     }
     colnames(RPKM) <- c("gene","intron","exon")
 
@@ -1091,5 +1135,143 @@ mRNAcontamination <- function(rpkm, raw=FALSE) {
     print(q)
     invisible(dev.off())
 }
+
+
+#' Plot the distribution of highest covered TSS density/gene body density ratios
+#'
+#' @param pi A single column containing the ratio of TSS densities/gene body
+#'           densities for the highest scoring TSSs
+#' @keywords pause index
+#' @export
+#' @examples
+#' data("pidx")
+#' plotPI(pi = "pidx")
+#' @export
+plotPI <- function(pi) {
+    if (exists(pi)) {
+        PI <- data.table(get(pi))
+    } else if (file.exists(pi)) {
+        PI <- fread(pi)
+    } else {
+        stop(paste0("FileExistsError: ", pi, " could not be found."))
+        quit(save = "no", status = 1, runLast = FALSE)
+    }
+    colnames(PI) <- c("pi")
+
+    name           <- basename(tools::file_path_sans_ext(pi))
+    numFields      <- 2
+    for(j in 1:numFields) name <- gsub("_[^_]*$", "", name)
+    sample_name <- paste(dirname(pi), name, sep="/")
+
+    q <- ggplot(data = PI, aes(x="", y=pi)) +
+            stat_boxplot(geom ='errorbar', width = 0.25) +
+            geom_boxplot(width = 0.25,
+                         outlier.color='red',
+                         outlier.shape=1) +
+            stat_summary(fun.y = "mean", geom = "point",
+                         shape = 1, size = 2) +
+            labs(x=name, y="each gene's pause index")
+
+    if (max(PI$pi) > 500) {
+        q <- q + scale_y_continuous(breaks = round(seq(min(PI$pi),
+                                              max(PI$pi),
+                                              by = 50), 0),
+                                    limits=c(0, max(PI$pi)))
+    } else if (max(PI$pi) > 100 & max(PI$pi) < 500) {
+        q <- q + scale_y_continuous(breaks = round(seq(min(PI$pi),
+                                              max(PI$pi),
+                                              by = 10), 0),
+                                    limits=c(0, max(PI$pi)))
+    } else {
+        q <- q + scale_y_continuous(breaks = round(seq(min(PI$pi),
+                                              max(PI$pi),
+                                              by = 5), 0),
+                                    limits=c(0, max(PI$pi)))
+    }
+    q <- q + coord_cartesian(ylim=c(0, ceiling(boxplot(PI$pi)$stats[5]))) +
+            theme_classic(base_size=14) +
+            theme(axis.line = element_line(size = 0.5)) +
+            theme(panel.grid.major = element_blank(),
+                  panel.grid.minor = element_blank(),
+                  aspect.ratio = 1,
+                  panel.border = element_rect(colour = "black",
+                                              fill=NA, size=0.5))
+    max_y  <- ceiling(boxplot(PI$pi)$stats[5])
+    label1 <- c(paste("'median'", ":", round(median(PI$pi), 2)),
+                paste("'mean'", ":", round(mean(PI$pi), 2)))
+    q <- q + annotate("text", x = 0.5, y = c(max_y, 0.95*max_y),
+                      hjust=0, vjust=1, label = label1, parse=TRUE)
+
+    # Save plot to pdf file
+    pdf(file=paste0(sample_name, "_pause_index.pdf"),
+        width= 7, height = 7, useDingbats=F)
+    print(q)
+    invisible(dev.off())
+         
+    # Save plot to png file
+    png(filename = paste0(sample_name, "_pause_index.png"),
+        width = 480, height = 480)
+    print(q)
+    invisible(dev.off())
+}
+
+
+#' Plot the distribution of adapter insertions
+#'
+#' @param input A cutadapt report
+#'
+#' @keywords cutadapt
+#' @export
+#' @examples
+#' data("cutadapt")
+#' plotCutadapt(input = "cutadapt")
+#' @export
+plotCutadapt <- function(input) {
+    if (exists(input)) {
+        report <- data.table(get(input))
+    } else if (file.exists(input)) {
+        report <- fread(input)
+    } else {
+        stop(paste0("FileExistsError: ", input, " could not be found."))
+        quit(save = "no", status = 1, runLast = FALSE)
+    }
+
+    name      <- basename(tools::file_path_sans_ext(input))
+    numFields <- 2
+    for(j in 1:numFields) name <- gsub("_[^_]*$", "", name)
+    sample_name <- paste(dirname(input), name, sep="/")
+    
+    # only keep sizes where the expected count represents less than 1% of 
+    # the actual count
+    report <- report[which(report$expect/report$count < 0.01),]
+    
+    # don't include size 0 insertions
+    report <- report[-nrow(report),]
+
+    q <- ggplot(report, aes(x=max(length)-length, y=count)) +
+            geom_point() +
+            labs(title=name, x="Size of insertion", y="Number of reads") +
+            theme_classic(base_size=14) +
+            theme(axis.line = element_line(size = 0.5),
+                  plot.title = element_text(hjust = 0.5),
+                  panel.grid.major = element_blank(),
+                  panel.grid.minor = element_blank(),
+                  aspect.ratio = 1,
+                  panel.border = element_rect(colour = "black",
+                                              fill=NA, size=0.5))
+
+    # Save plot to pdf file
+    pdf(file=paste0(sample_name, "_adapter_insertion_distribution.pdf"),
+        width= 7, height = 7, useDingbats=F)
+    print(q)
+    invisible(dev.off())
+         
+    # Save plot to png file
+    png(filename = paste0(sample_name, "_adapter_insertion_distribution.png"),
+        width = 480, height = 480)
+    print(q)
+    invisible(dev.off())
+}
+
 
 ################################################################################